# Add caching proxy with major performance improvements (92 MB/s)