--- a/.gitignore
+++ b/.gitignore
@@ -24,3 +24,4 @@ config.toml
 
 flamegraph.svg
 perf.data*
+cache-config.toml