@mjc 2025-10-07
Re: Arc removal thread safety concern

The Arc wrapper was redundant. Here's why this is safe:

1. `DeadpoolConnectionProvider` implements `Clone` and contains `managed::Pool<TcpManager>`
2. Deadpool's `Pool` type is Arc-based internally (standard for connection pools that need to be shared across threads)
3. When `NntpProxy` is cloned (e.g., for each client task), the `Vec<DeadpoolConnectionProvider>` is cloned
4. Cloning `DeadpoolConnectionProvider` just clones the Arc inside its Pool - all clones share the same underlying connection pool

So we have:
- **Before**: `Arc<Vec<DeadpoolConnectionProvider>>` - Arc of Vec of Pool (where Pool is already Arc-based)
- **After**: `Vec<DeadpoolConnectionProvider>` - Vec of Pool (Pool is still Arc-based)

The redundant outer Arc has been removed. Thread safety is maintained because the actual connection pools are shared via the Arc inside `Pool`, not via the Vec container.