@mjc 2025-10-07
Re: Magic number 65536

Fixed! Replaced with `STREAMING_CHUNK_SIZE` constant from the unified constants module (`src/constants.rs`). All constants have been consolidated into a single location as part of this PR.