@mjc 2025-10-07
Re: Cloning in PoolPrewarmer constructor (updated)

Actually, you were right to point this out - but the real issue was deeper. The `PoolPrewarmer` struct itself was unnecessary overhead:

**Root cause**: `PoolPrewarmer` was stored in `NntpProxy`, which is cloned for every client connection. But prewarming only happens once at startup.

**Solution**: Removed the `PoolPrewarmer` struct entirely and replaced it with a simple `prewarm_pools()` function that:
- Takes slices directly (no cloning needed)
- Is only called once at startup
- Does not need state tracking (removed pointless AtomicBool)
- Is not stored anywhere (no cloning overhead)

This is cleaner than adding Arc wrappers to a struct that should not exist in the first place. Thanks for catching this.